+++ cucumber-junit-platform-engine/src/main/java/io/cucumber/junit/platform/engine/CucumberTestEngine.java
@@ -15,6 +15,7 @@ import org.junit.platform.engine.support.hierarchical.HierarchicalTestEngine;
 import org.junit.platform.engine.support.hierarchical.HierarchicalTestExecutorService;
 
 import static io.cucumber.junit.platform.engine.Constants.FEATURES_PROPERTY_NAME;
+import static io.cucumber.junit.platform.engine.Constants.JUNIT_PLATFORM_DISCOVERY_AS_ROOT_ENGINE_PROPERTY_NAME;
 import static io.cucumber.junit.platform.engine.Constants.PARALLEL_CONFIG_PREFIX;
 import static org.junit.platform.engine.support.discovery.DiscoveryIssueReporter.deduplicating;
 import static org.junit.platform.engine.support.discovery.DiscoveryIssueReporter.forwarding;
@@ -44,8 +45,9 @@ public final class CucumberTestEngine extends HierarchicalTestEngine<CucumberEng
 
     @Override
     public TestDescriptor discover(EngineDiscoveryRequest discoveryRequest, UniqueId uniqueId) {
-        TestSource testSource = createEngineTestSource(discoveryRequest);
-        CucumberConfiguration configuration = new CucumberConfiguration(discoveryRequest.getConfigurationParameters());
+        ConfigurationParameters configurationParameters = discoveryRequest.getConfigurationParameters();
+        TestSource testSource = createEngineTestSource(configurationParameters);
+        CucumberConfiguration configuration = new CucumberConfiguration(configurationParameters);
         CucumberEngineDescriptor engineDescriptor = new CucumberEngineDescriptor(uniqueId, configuration, testSource);
 
         DiscoveryIssueReporter issueReporter = deduplicating(forwarding( //
@@ -53,19 +55,35 @@ public final class CucumberTestEngine extends HierarchicalTestEngine<CucumberEng
             engineDescriptor.getUniqueId() //
         ));
 
+        // Early out if Cucumber is the root engine and discovery has been
+        // explicitly disabled. Workaround for:
+        // https://github.com/sbt/sbt-jupiter-interface/issues/142
+        if (!supportsDiscoveryAsRootEngine(configurationParameters) && isRootEngine(uniqueId)) {
+            return engineDescriptor;
+        }
+
         FeaturesPropertyResolver resolver = new FeaturesPropertyResolver(new DiscoverySelectorResolver());
         resolver.resolveSelectors(discoveryRequest, engineDescriptor, issueReporter);
         return engineDescriptor;
     }
 
-    private static TestSource createEngineTestSource(EngineDiscoveryRequest discoveryRequest) {
+    private static boolean supportsDiscoveryAsRootEngine(ConfigurationParameters configurationParameters) {
+        return configurationParameters.getBoolean(JUNIT_PLATFORM_DISCOVERY_AS_ROOT_ENGINE_PROPERTY_NAME)
+                .orElse(true);
+    }
+
+    private boolean isRootEngine(UniqueId uniqueId) {
+        UniqueId cucumberRootEngineId = UniqueId.forEngine(getId());
+        return uniqueId.hasPrefix(cucumberRootEngineId);
+    }
+
+    private static TestSource createEngineTestSource(ConfigurationParameters configurationParameters) {
         // Workaround. Test Engines do not normally have test source.
         // Maven does not count tests that do not have a ClassSource somewhere
         // in the test descriptor tree.
         // Gradle will report all tests as coming from an "Unknown Class"
         // See: https://github.com/cucumber/cucumber-jvm/pull/2498
-        ConfigurationParameters configuration = discoveryRequest.getConfigurationParameters();
-        if (configuration.get(FEATURES_PROPERTY_NAME).isPresent()) {
+        if (configurationParameters.get(FEATURES_PROPERTY_NAME).isPresent()) {
             return ClassSource.from(CucumberTestEngine.class);
         }
         return null;